most stuff cleaned up and working note, the default config assumes data and dsrc folders are set up in the project root directrory tres, tocs and overwrite files should be put into data/SwgClient some tools not working/breaking due to missing server/data files or cause of stoneage perforce dependancies, some require flex/bison/perl to be installed and in the system path

+++ src/external/3rd/library/platform/projects/Session/LoginAPI/Messages.cpp
@@ -24,7 +24,7 @@ namespace Base
 
     void  put(ByteStream & target, const LoginAPI::UsageLimit & source)
     {
-	unsigned next = source.GetNextAllowance();
+	    unsigned next = (unsigned int)source.GetNextAllowance();
         put(target, source.GetType());
         put(target, source.GetAllowance());
         put(target, next);
